moved api in timeline FFI
fix clippy warning
